refactor: clean up MultiSelect imports and exports
