+++ packages/react/src/components/ComposedModal/ComposedModal.featureflag.stories.js
@@ -1,31 +1,19 @@
 /**
- * Copyright IBM Corp. 2016, 2023
+ * Copyright IBM Corp. 2016, 2025
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
 
-import React, { useState, useRef } from 'react';
-import ReactDOM from 'react-dom';
+import React, { useState } from 'react';
 import ComposedModal, { ModalBody } from './ComposedModal';
 import { ModalHeader } from './ModalHeader';
 import { ModalFooter } from './ModalFooter';
-import MultiSelect from '../MultiSelect';
-import Dropdown from '../Dropdown';
 import Select from '../Select';
 import SelectItem from '../SelectItem';
 import TextInput from '../TextInput';
 import Button from '../Button';
-import {
-  StructuredListWrapper,
-  StructuredListHead,
-  StructuredListBody,
-  StructuredListRow,
-  StructuredListCell,
-} from '../StructuredList';
-import mdx from './ComposedModal.featureflag.mdx';
 import { WithFeatureFlags } from '../../../.storybook/templates/WithFeatureFlags';
-import { title } from 'process';
 
 export default {
   title: 'Components/ComposedModal/Feature Flag',
