+++ packages/react/src/components/FluidMultiSelect/FluidFilterableMultiSelect.tsx
@@ -1,5 +1,5 @@
 /**
- * Copyright IBM Corp. 2022
+ * Copyright IBM Corp. 2022, 2025
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -8,17 +8,13 @@
 import PropTypes from 'prop-types';
 import React, { ForwardedRef, FunctionComponent, RefObject } from 'react';
 import classnames from 'classnames';
-import { FilterableMultiSelect } from '../FilterableMultiSelect';
-import MultiSelect from '../MultiSelect';
 import { usePrefix } from '../../internal/usePrefix';
 import { FormContext } from '../FluidForm/FormContext';
+import type { UseComboboxProps, UseMultipleSelectionProps } from 'downshift';
 import {
-  UseComboboxProps,
-  UseMultipleSelectionProps,
-  UseSelectProps,
-} from 'downshift';
-import { MultiSelectProps } from '../MultiSelect/MultiSelect';
-import { FilterableMultiSelectProps } from '../MultiSelect/FilterableMultiSelect';
+  FilterableMultiSelect,
+  type FilterableMultiSelectProps,
+} from '../MultiSelect';
 
 interface OnChangeData<ItemType> {
   selectedItems: ItemType[] | null;
