+++ packages/react/src/components/MultiSelect/FilterableMultiSelect.tsx
@@ -32,6 +32,7 @@ import React, {
   ReactElement,
   useLayoutEffect,
   useMemo,
+  forwardRef,
 } from 'react';
 import { defaultFilterItems } from '../ComboBox/tools/filter';
 import {
@@ -312,7 +313,7 @@ export interface FilterableMultiSelectProps<ItemType>
   warnText?: ReactNode;
 }
 
-const FilterableMultiSelect = React.forwardRef(function FilterableMultiSelect<
+export const FilterableMultiSelect = forwardRef(function FilterableMultiSelect<
   ItemType,
 >(
   {
@@ -1016,6 +1017,7 @@ const FilterableMultiSelect = React.forwardRef(function FilterableMultiSelect<
   displayName?: any;
 };
 
+FilterableMultiSelect.displayName = 'FilterableMultiSelect';
 FilterableMultiSelect.propTypes = {
   /**
    * Deprecated, aria-label is no longer needed
@@ -1213,5 +1215,3 @@ FilterableMultiSelect.propTypes = {
    */
   warnText: PropTypes.node,
 };
-
-export default FilterableMultiSelect;
