+++ packages/react/src/index.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright IBM Corp. 2016, 2023
+ * Copyright IBM Corp. 2016, 2025
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -34,7 +34,6 @@ export * from './components/Dropdown';
 export * from './components/ErrorBoundary';
 export * from './components/ExpandableSearch';
 export * from './components/FileUploader';
-export * from './components/FilterableMultiSelect';
 export * from './components/FluidForm';
 export * from './components/Form';
 export * from './components/FormGroup';
